modularize devtools and custom code
