+++ devtools/timelineModel/counterGraph/calculator.ts
