+++ devtools/timelineModel/counterGraph/counter.ts
