+++ devtools/timelineModel/timelineFrame/layerPaintEvent.ts
