+++ devtools/tracingModel/namedObject.ts
