+++ devtools/tracingModel/objectSnapshot.ts
